+++ tests/grpc/client/query/bank.go
@@ -1,4 +1,4 @@
-package client
+package query
 
 import (
 	"context"
