+++ tests/grpc/client/query/client.go
@@ -1,4 +1,4 @@
-package client
+package query
 
 import (
 	"time"
