+++ tests/grpc/client/query/grpc.go
@@ -1,4 +1,4 @@
-package client
+package query
 
 import (
 	"context"
