+++ tests/grpc/client/query/oracle.go
@@ -1,4 +1,4 @@
-package client
+package query
 
 import (
 	"context"
