+++ tests/grpc/client/tx/bank.go
@@ -1,4 +1,4 @@
-package client
+package tx
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
