+++ tests/grpc/client/tx/client.go
@@ -1,4 +1,4 @@
-package client
+package tx
 
 import (
 	"os"
