+++ tests/grpc/client/tx/gov.go
@@ -1,4 +1,4 @@
-package client
+package tx
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
