+++ tests/grpc/client/tx/sign.go
@@ -1,4 +1,4 @@
-package client
+package tx
 
 import (
 	"github.com/cosmos/cosmos-sdk/client"
