+++ tests/grpc/client/umee_client.go
@@ -1,12 +1,17 @@
 package client
 
+import (
+	"github.com/umee-network/umee/v4/tests/grpc/client/query"
+	"github.com/umee-network/umee/v4/tests/grpc/client/tx"
+)
+
 // UmeeClient is a helper for initializing a keychain, a cosmos-sdk client context,
 // and sending transactions/queries to a specific Umee node
 // It also starts up a websocket connection to track the current block height and
 // uses the block height to ensure transactions happen within a certain window.
 type UmeeClient struct {
-	QueryClient *QueryClient
-	TxClient    *TxClient
+	QueryClient *query.QueryClient
+	TxClient    *tx.TxClient
 }
 
 func NewUmeeClient(
@@ -17,7 +22,7 @@ func NewUmeeClient(
 	accountMnemonic string,
 ) (uc *UmeeClient, err error) {
 	uc = &UmeeClient{}
-	uc.QueryClient = NewQueryClient(grpcEndpoint)
-	uc.TxClient, err = NewTxClient(chainID, tmrpcEndpoint, accountName, accountMnemonic)
+	uc.QueryClient = query.NewQueryClient(grpcEndpoint)
+	uc.TxClient, err = tx.NewTxClient(chainID, tmrpcEndpoint, accountName, accountMnemonic)
 	return uc, err
 }
